Merged PR 92245: Add support for logging to stdout
Added a new stream to bunyanLogger that writes to stdout. The stdout stream is only created if logToStdout, a new parameter of createBunyanSeqLogger, is set to true. It is set to true by default, so the new default behavior is to log to the console as well as SEQ.

Related work items: #321899