don't download unnecessary changesets in the transformer and refactor processChanges test (#4949)
previously, calling `processChanges` would download from the changeset to start processing, all the way to the latest. This is unnecessary, especially since we explicitly support processing non-latest changes by requiring the consumer to checkout their source iModel at an earlier changeset (at the last changeset to process). This meant in a common case we would unnecessary download excess changesets. This PR fixes that, and adds a test to ensure we don't download past the source's changeset.

I also refactored the large `processChanges` test, storing the series of events as a "timeline", a javascript object containing which events (e.g. creation of new iModels, updates to content, branching, synchronization) to occur through time during the test.



Co-authored-by: Michael Belousov <MichaelBelousov@users.noreply.github.com>