Add branch name to reporter (#67)
List branch name in reporter, so we can differentiate where the tests
were ran, and compare performance based on this.
Currently I attached the branch as field in the json `info` section of
the db. However, I also attached it to the imodel-name because when
uploading to the db, and getting processed in power BI this should allow
us to compare the run times very easily. I don't really like this
solution, but I couldn't figure out a way to use the `info` fields in
Power BI.

Also, I changed the logger setup, but I'm not logging the entities being
processed, so I think something is wrong.

---------

Co-authored-by: Daniel Rodriguez <DanRod1999@users.noreply.github.com>