Read changesets directly instead of applying them to determine whats changed (#135)
Keeping the PR as draft at the moment, because waiting on a change from
Affan to the changesetECAdaptor to not error out when there is no
ECClassId and instead fall back to the ecclassid of the root class.

The above scenario occurs when the element is no longer present in the
database (because it was deleted in some future changeset that isn't
currently being processed) so it can't be queried for its ecclassid.

EDIT: In 4.3.3 there is a different issue with classIdPresentInChange.
See below PR
https://github.com/iTwin/itwinjs-core/pull/6341

---------

Co-authored-by: Michael Belousov <mike.belousov@bentley.com>