Federation guid optimization (#31)
Instead of using External Source Aspects for transformations, keep track
of the previously-transformed source changeset id in the version field
of only one external source aspect in the target scoping element in the
target, and use changeset data to determine if an iModel has been
changed when retransforming. This allows us to skip a lot of updates
like we did with the External Source Aspect scheme, but without writing
a new aspect on transform relation initialization (e.g. branch forking)

---------

Co-authored-by: imodeljs-admin <imodeljs-admin@users.noreply.github.com>
Co-authored-by: Vilius Ruskys <36619139+ViliusRuskys@users.noreply.github.com>
Co-authored-by: Kyrylo Volotovskyi <130470116+KyryloVolotovskyi@users.noreply.github.com>
Co-authored-by: BENTLEY\Kyrylo.Volotovskyi <Kyrylo.Volotovskyi@bentley.com>
Co-authored-by: Daniel Rodriguez <44824788+DanRod1999@users.noreply.github.com>
Co-authored-by: Daniel Rodriguez <DanRod1999@users.noreply.github.com>
Co-authored-by: nick.tessier <22119573+nick4598@users.noreply.github.com>
Co-authored-by: Mindaugas Dirginčius <25615181+mindaugasdirg@users.noreply.github.com>
Co-authored-by: Jiarui Zhang <91682445+jiaruiz717@users.noreply.github.com>