Revert "save and push changes after transformer is complete (#153)" (#185)
Leaving the responsibility up to the consumer. We already require them
to download the iModels and acquire locks themselves, so we should also
let them push changes as pushing changes releases locks.

There have also been mentions of some transforms intentionally inserting
data that isn't meant to be transformed back over, which would not be
possible if we pushed changes for consumers, because then they wouldn't
be able to add that data as part of the pending changeset indices.