No longer track partially commited element references (#206)
Tracking each unresolved reference to be able to resolve it as soon as
possible has a big impact on memory consumption when there are large
amounts of pending references.
One such case is multiple display styles with hundreds of thousands of
excluded elements.

The transformer now only keeps track of element and aspect ids that were
only partially completed. These elements and aspects will be reprocessed
once all the elements have been processed by the transformer. The only
unresolvable references at that point will be the elements that were
intentionally filtered out by the transformer or the invalid ones.

---------

Co-authored-by: Nick Tessier <22119573+nick4598@users.noreply.github.com>