Allow Performing Transforms with additionalTransforms (#265)
If src and target imodel's have differing local Helmert transforms,
transform target elements positions to maintain proper local positions.

The target imodel "additionalTransform" will always be maintained, even
if it is null. The source imodel when copied into the target will have a
combination of both helmert transforms applied to its elements
placements. This way the result will place the elements in a position
that when the targets "additionalTransform" is applied it will be in the
correct geographic position. If the target's "additionalTransform" is
null, then the we are simply applying the src's helmert transform
permanently to each elements position instead of just at render time. If
the src and target have different non null transforms, then the
resulting transform applied to the src elements will also negate the
helmert transform applied by the target model.

---------

Co-authored-by: Daniel Rodriguez <DanRod1999@users.noreply.github.com>